Crate quaternion

A simple and type agnostic quaternion math library designed for reexporting

Functions

add
Adds two quaternions.

axis_angle
Construct a quaternion for the given angle (in radians) about the given axis. Axis must be a unit vector.

conj
Takes the quaternion conjugate.

euler_angles
Construct a quaternion representing the given euler angle rotations (in radians)

id
Constructs identity quaternion.

len
Computes the length of a quaternion.

mul
Multiplies two quaternions.

rotate_vector
Rotate the given vector using the given quaternion

square_len
Computes the square length of a quaternion.

Type Definitions

Quaternion
Quaternion type alias.