Function quaternion::rotate_a_to_b

pub fn rotate_a_to_b<T>(a: Vector3<T>, b: Vector3<T>, t: T) -> Quaternion<T> where
    T: Float + FloatConst, 

位置ベクトルaを 位置ベクトルbと同じ場所へ最短距離で回転させるVersorを求める．

零ベクトルを入力した場合は，恒等四元数を返す．

Find a versor to rotate from vector a to b. If you enter a zero vector, it returns an identity quaternion.

Range of rotation parameter: 0 <= t <= 1